refactor: set default exchange currency based on locale
